requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
yfinance>=0.1.63
requests>=2.26.0
python-dotenv>=0.19.0
asyncio>=3.4.3
websockets>=9.1
ccxt>=1.60.0
ta>=0.7.0
plotly>=5.3.0